requirements.txt
gradio==6.3.0
gunicorn
uvicorn[standard]
pydantic
python-multipart
sse-starlette
google-genai==1.57.0
playwright==1.57.0
browserbase==1.4.0
stagehand==0.5.9